requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
asyncio-dgram==2.1.2
attrs==23.1.0
charset-normalizer==3.3.0
discord==2.3.2
discord.py==2.3.2
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
mcstatus==11.0.1
multidict==6.0.4
python-dotenv==1.0.0
yarl==1.9.2
